+++ CHANGELOG.md
@@ -6,6 +6,16 @@
 ### BREAKING CHANGES
 -->
 
+<a name="1.0.2"></a>
+
+## 1.0.2 (2017-04-20)
+
+### Features
+ * Interface `ExportReadable` to provide String representation of objects or enumerations
+ * Method `buildDOC` without the need for a file name
+ * No EJB components anymore, just plain CDI
+
+
 <a name="1.0.1"></a>
 
 ## 1.0.1 (2017-04-05)